pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dy-oppty-agent"
version = "0.1.0"
description = "Douyin Shop Opportunity Center Agent scaffold"
readme = "README.md"
requires-python = ">=3.11,<3.12"
authors = [{ name = "dy-oppty-agent" }]
dependencies = [
  "playwright>=1.48.0",
  "typer>=0.12.0",
  "pydantic>=2.9.0",
  "PyYAML>=6.0.2",
  "pandas>=2.2.3",
  "openpyxl>=3.1.5",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.9",
  "pytest>=8.3.3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["oppty_agent*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
